+++ backend/src/apiserver/server/job_server_test.go
@@ -70,7 +70,7 @@ var (
 func TestValidateApiJob(t *testing.T) {
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	err := server.validateCreateJobRequest(&api.CreateJobRequest{Job: commonApiJob})
 	assert.Nil(t, err)
 }
@@ -78,7 +78,7 @@ func TestValidateApiJob(t *testing.T) {
 func TestValidateApiJob_WithPipelineVersion(t *testing.T) {
 	clients, manager, _ := initWithExperimentAndPipelineVersion(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	apiJob := &api.Job{
 		Name:           "job1",
 		Enabled:        true,
@@ -97,7 +97,7 @@ func TestValidateApiJob_WithPipelineVersion(t *testing.T) {
 func TestValidateApiJob_ValidateNoExperimentResourceReferenceSucceeds(t *testing.T) {
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	apiJob := &api.Job{
 		Name:           "job1",
 		Enabled:        true,
@@ -120,7 +120,7 @@ func TestValidateApiJob_ValidateNoExperimentResourceReferenceSucceeds(t *testing
 func TestValidateApiJob_ValidatePipelineSpecFailed(t *testing.T) {
 	clients, manager, experiment := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	apiJob := &api.Job{
 		Name:           "job1",
 		Enabled:        true,
@@ -146,7 +146,7 @@ func TestValidateApiJob_ValidatePipelineSpecFailed(t *testing.T) {
 func TestValidateApiJob_NoValidPipelineSpecOrPipelineVersion(t *testing.T) {
 	clients, manager, _ := initWithExperimentAndPipelineVersion(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	apiJob := &api.Job{
 		Name:           "job1",
 		Enabled:        true,
@@ -166,7 +166,7 @@ func TestValidateApiJob_NoValidPipelineSpecOrPipelineVersion(t *testing.T) {
 func TestValidateApiJob_InvalidCron(t *testing.T) {
 	clients, manager, experiment := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	apiJob := &api.Job{
 		Name:           "job1",
 		Enabled:        true,
@@ -192,7 +192,7 @@ func TestValidateApiJob_InvalidCron(t *testing.T) {
 func TestValidateApiJob_MaxConcurrencyOutOfRange(t *testing.T) {
 	clients, manager, experiment := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	apiJob := &api.Job{
 		Name:           "job1",
 		Enabled:        true,
@@ -218,7 +218,7 @@ func TestValidateApiJob_MaxConcurrencyOutOfRange(t *testing.T) {
 func TestValidateApiJob_NegativeIntervalSecond(t *testing.T) {
 	clients, manager, experiment := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	apiJob := &api.Job{
 		Name:           "job1",
 		Enabled:        true,
@@ -243,7 +243,7 @@ func TestValidateApiJob_NegativeIntervalSecond(t *testing.T) {
 func TestCreateJob(t *testing.T) {
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	job, err := server.CreateJob(nil, &api.CreateJobRequest{Job: commonApiJob})
 	assert.Nil(t, err)
 	assert.Equal(t, commonExpectedJob, job)
@@ -258,7 +258,7 @@ func TestCreateJob_Unauthorized(t *testing.T) {
 
 	clients, manager, _ := initWithExperiment_KFAM_Unauthorized(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	_, err := server.CreateJob(ctx, &api.CreateJobRequest{Job: commonApiJob})
 	assert.NotNil(t, err)
 	assert.Contains(t, err.Error(), "Unauthorized access")
@@ -273,13 +273,13 @@ func TestGetJob_Unauthorized(t *testing.T) {
 
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	job, err := server.CreateJob(ctx, &api.CreateJobRequest{Job: commonApiJob})
 	assert.Nil(t, err)
 
 	clients.KfamClientFake = client.NewFakeKFAMClientUnauthorized()
 	manager = resource.NewResourceManager(clients)
-	server = NewJobServer(manager)
+	server = NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 
 	_, err = server.GetJob(ctx, &api.GetJobRequest{Id: job.Id})
 	assert.NotNil(t, err)
@@ -295,7 +295,7 @@ func TestGetJob_Multiuser(t *testing.T) {
 
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	createdJob, err := server.CreateJob(ctx, &api.CreateJobRequest{Job: commonApiJob})
 	assert.Nil(t, err)
 
@@ -313,7 +313,7 @@ func TestListJobs_Unauthorized(t *testing.T) {
 
 	clients, manager, experiment := initWithExperiment_KFAM_Unauthorized(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	_, err := server.ListJobs(ctx, &api.ListJobsRequest{
 		ResourceReferenceKey: &api.ResourceKey{
 			Type: api.ResourceType_EXPERIMENT,
@@ -342,7 +342,7 @@ func TestListJobs_Multiuser(t *testing.T) {
 
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	_, err := server.CreateJob(ctx, &api.CreateJobRequest{Job: commonApiJob})
 	assert.Nil(t, err)
 
@@ -442,13 +442,13 @@ func TestEnableJob_Unauthorized(t *testing.T) {
 
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	job, err := server.CreateJob(ctx, &api.CreateJobRequest{Job: commonApiJob})
 	assert.Nil(t, err)
 
 	clients.KfamClientFake = client.NewFakeKFAMClientUnauthorized()
 	manager = resource.NewResourceManager(clients)
-	server = NewJobServer(manager)
+	server = NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 
 	_, err = server.EnableJob(ctx, &api.EnableJobRequest{Id: job.Id})
 	assert.NotNil(t, err)
@@ -464,7 +464,7 @@ func TestEnableJob_Multiuser(t *testing.T) {
 
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 
 	job, err := server.CreateJob(ctx, &api.CreateJobRequest{Job: commonApiJob})
 	assert.Nil(t, err)
@@ -482,13 +482,13 @@ func TestDisableJob_Unauthorized(t *testing.T) {
 
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 	job, err := server.CreateJob(ctx, &api.CreateJobRequest{Job: commonApiJob})
 	assert.Nil(t, err)
 
 	clients.KfamClientFake = client.NewFakeKFAMClientUnauthorized()
 	manager = resource.NewResourceManager(clients)
-	server = NewJobServer(manager)
+	server = NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 
 	_, err = server.DisableJob(ctx, &api.DisableJobRequest{Id: job.Id})
 	assert.NotNil(t, err)
@@ -504,7 +504,7 @@ func TestDisableJob_Multiuser(t *testing.T) {
 
 	clients, manager, _ := initWithExperiment(t)
 	defer clients.Close()
-	server := NewJobServer(manager)
+	server := NewJobServer(manager, &JobServerOptions{CollectMetrics: false})
 
 	job, err := server.CreateJob(ctx, &api.CreateJobRequest{Job: commonApiJob})
 	assert.Nil(t, err)