+++ components/aws/sagemaker/tests/integration_tests/component_tests/test_train_component.py
@@ -16,6 +16,7 @@ from utils import argo_utils
         pytest.param("resources/config/fsx-mnist-training", marks=pytest.mark.fsx_test),
         "resources/config/spot-sample-pipeline-training",
         "resources/config/assume-role-training",
+        "resources/config/xgboost-mnist-trainingjob-debugger",
     ],
 )
 def test_trainingjob(