+++ components/aws/sagemaker/Changelog.md
@@ -4,6 +4,17 @@ The version of the AWS SageMaker Components is determined by the docker image ta
 Repository:  https://hub.docker.com/repository/docker/amazon/aws-sagemaker-kfp-components
 
 ---------------------------------------------
+**Change log for version 0.9.0**
+- Add functionality to Update SageMaker Endpoint for Deploy component
+
+> Pull requests : [#4424](https://github.com/kubeflow/pipelines/pull/4424/)
+
+**Change log for version 0.8.0**
+- Add functionality to configure SageMaker Debugger for Training component
+
+> Pull requests : [#4283](https://github.com/kubeflow/pipelines/pull/4283/)
+
+
 **Change log for version 0.7.0**
 - Add functionality to assume role when sending SageMaker requests
 
@@ -29,7 +40,7 @@ Repository:  https://hub.docker.com/repository/docker/amazon/aws-sagemaker-kfp-c
 
 
 **Change log for version 0.5.1**
-- Update region support for GroudTruth component
+- Update region support for GroundTruth component
 - Make `label_category_config` an optional parameter in Ground Truth component
 
 > Pull requests : [#3932](https://github.com/kubeflow/pipelines/pull/3932)